ApisecCaptureAnalyzerService-main/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9

# Data processing and ML - using newer versions for Python 3.13 compatibility
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.5.0

# Utilities
python-dateutil>=2.9.0
jsonschema>=4.23.0
pyyaml>=6.0.0

# Storage (placeholders for now)
boto3>=1.35.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Development
black>=24.0.0
ruff>=0.6.0

